+++ src/RejectedPromise.php
@@ -4,7 +4,7 @@ namespace React\Promise;
 
 use AsyncInterop\Promise as AsyncInteropPromise;
 
-class RejectedPromise implements ExtendedPromiseInterface, CancellablePromiseInterface
+class RejectedPromise implements PromiseInterface
 {
     private $reason;
 
@@ -49,7 +49,7 @@ class RejectedPromise implements ExtendedPromiseInterface, CancellablePromiseInt
                 throw UnhandledRejectionException::resolve($result->reason);
             }
 
-            if ($result instanceof ExtendedPromiseInterface) {
+            if ($result instanceof PromiseInterface) {
                 $result->done();
             }
         });